+++ packages/jest-snapshot/package.json
@@ -35,8 +35,9 @@
     "jest-message-util": "workspace:*",
     "jest-util": "workspace:*",
     "natural-compare": "^1.4.0",
-    "pretty-format": "workspace:*",
-    "semver": "^7.5.3"
+    "pretty-format": "workspace:^",
+    "semver": "^7.5.3",
+    "synckit": "^0.8.5"
   },
   "devDependencies": {
     "@babel/preset-flow": "^7.7.2",
@@ -46,11 +47,12 @@
     "@types/babel__core": "^7.1.14",
     "@types/graceful-fs": "^4.1.3",
     "@types/natural-compare": "^1.4.0",
-    "@types/prettier": "^2.1.5",
+    "@types/prettier-v2": "npm:@types/prettier@^2.1.5",
     "@types/semver": "^7.1.0",
     "ansi-regex": "^5.0.1",
     "ansi-styles": "^5.0.0",
-    "prettier": "^2.1.1",
+    "prettier": "^3.0.3",
+    "prettier-v2": "npm:prettier@^2.1.5",
     "tsd-lite": "^0.8.0"
   },
   "engines": {
